refactor(ggml): begin decoupling the old formats
