+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "rust-analyzer.cargo.features": []
+  "rust-analyzer.cargo.features": ["pre-gguf-formats"]
 }
