+++ crates/ggml/src/util.rs
@@ -7,13 +7,13 @@ use std::{
 
 /// Helper struct that wraps the magic number of a file format,
 /// so that it can be printed in a human-readable format.
-pub struct FormatMagic(pub [u8; 4]);
-impl fmt::Display for FormatMagic {
+pub struct FileMagic(pub [u8; 4]);
+impl fmt::Display for FileMagic {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{:x?} ({})", self.0, String::from_utf8_lossy(&self.0))
     }
 }
-impl fmt::Debug for FormatMagic {
+impl fmt::Debug for FileMagic {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> std::fmt::Result {
         fmt::Display::fmt(self, f)
     }
