+++ crates/llm-base/Cargo.toml
@@ -11,7 +11,7 @@ readme = "../../README.md"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ggml = { path = "../ggml", version = "0.2.0-dev" }
+ggml = { path = "../ggml", version = "0.2.0-dev", features = ["pre-gguf-formats"] }
 
 bytemuck = { workspace = true }
 rand = { workspace = true }
